src/requirements.txt
pyjwt<=2.10.1
aws_secretsmanager_caching<=1.3.1
requests>=2.10.0